pyproject.toml
[project]
name = "campus-event-management"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiosmtplib>=5.1.0",
    "authlib>=1.6.7",
    "bcrypt>=4.0.0",
    "fastapi>=0.128.7",
    "google-auth>=2.48.0",
    "groq>=0.4.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
